Addmessages

+++ Controller/CommentController.php
@@ -58,7 +58,8 @@ class CommentController extends Controller
     {
         $auth = getAuth();
         if($auth){
-            if($auth->id == $comment->userID){
+            $comment = $this->commentManager->getOne($id)->userID;
+            if($auth->id == $comment){
                 $comment = new \stdClass();
                 $comment->content = $_POST->content;
                 $comment->userID = $auth->id;
@@ -71,16 +72,27 @@ class CommentController extends Controller
                 }
             }else{
                 $this->JSONMessage("Vous n'avez pas les droits pour modifier cet commentaire");
+            }else{
+                $this->JSONMessage("Vous n'êtes pas authentifié.");
             }
         }
     }
 
     function delete($id)
-    {
-        if ($this->commentManager->delete($id)) {
-            $this->JSONMessage("commentaire supprimé");
-        } else {
-            $this->JSONMessage("commentaire non trouvé");
+    {   $auth = getAuth();
+        if($auth){
+            $comment = $this->commentManager->getOne($id)->userID;
+            if($auth->id == $comment){
+                if ($this->commentManager->delete($id)) {
+                    $this->JSONMessage("commentaire supprimé");
+                } else {
+                    $this->JSONMessage("commentaire non trouvé");
+                }
+            }else{
+                $this->JSONMessage("Vous n'avez pas les droits pour supprimer cet commentaire");
+            }
+        }else{
+            $this->JSONMessage("Vous n'êtes pas authentifié.");
         }
     }   
 }
